backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0
PyJWT==2.8.0

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Celery for background tasks
celery==5.3.6
flower==2.0.1

# Image Processing
Pillow==10.2.0
qrcode[pil]==7.4.2

# Geolocation
geopy==2.4.1

# File Storage (optional - use if deploying to cloud)
boto3==1.34.34

# Email (optional)
fastapi-mail==1.4.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3.post1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
